+++ app/models/federails/activity.rb
@@ -31,6 +31,17 @@ module Federails
     serialize :bto, coder: YAML
     serialize :bcc, coder: YAML
     serialize :audience, coder: YAML
+    serialize :result, coder: YAML
+    serialize :instrument, coder: YAML
+
+    # Sends an Undo activity to reverse this one.
+    #
+    # Used for unfollowing, unliking, unboosting, etc.
+    #
+    # @return [Federails::Activity] the new Undo activity
+    def undo!
+      Activity.create! actor: actor, action: 'Undo', entity: self, to: to, cc: cc
+    end
 
     private