+++ app/models/federails/following.rb
@@ -92,7 +92,7 @@ module Federails
 
     #: () -> Federails::Activity
     def destroy_activity
-      Activity.create! actor: actor, action: 'Undo', entity: follow_activity, to: [target_actor.federated_url]
+      follow_activity&.undo!
     end
   end
 end