+++ db/migrate/20260421170640_add_result_and_instrument_to_federails_activities.rb
@@ -1,0 +1,6 @@
+class AddResultAndInstrumentToFederailsActivities < ActiveRecord::Migration[7.2]
+  def change
+    add_column :federails_activities, :result, :string
+    add_column :federails_activities, :instrument, :string
+  end
+end